+++ projects/vscode/src/config.ts
@@ -320,14 +320,14 @@ class TrumboEndpoint {
 					apiBaseUrl: "http://localhost:7777",
 					mcpBaseUrl: "http://localhost:8787/v1/mcp",
 				}
-	default:
-		return {
-			environment: Environment.production,
-			appBaseUrl: "https://platform.trumbo.dev",
-			apiBaseUrl: "https://platform.trumbo.dev",
-			mcpBaseUrl: "https://platform.trumbo.dev/v1/mcp",
+			default:
+				return {
+					environment: Environment.production,
+					appBaseUrl: "https://platform.trumbo.dev",
+					apiBaseUrl: "https://api.trumbo.dev",
+					mcpBaseUrl: "https://api.trumbo.dev/v1/mcp",
+				}
 		}
-	}
 	}
 }